+++ mandelbrot_four_split.c
@@ -10,7 +10,7 @@
 typedef struct
 {
     Queue *q;
-    uint16_t *result;
+    double *result;
     Zoom *zoom;
     uint16_t max_iterations;
     int *sem;
@@ -23,13 +23,13 @@ static int rectangle_size(const Rectangle rectangle)
     return (rectangle.br.x - rectangle.tl.x) * (rectangle.br.y - rectangle.tl.y);
 }
 
-static bool rect_is_black(const Rectangle rect, const uint16_t *result, const Zoom *zoom, const uint16_t max_iterations)
+static bool rect_is_black(const Rectangle rect, const double *result, const Zoom *zoom, const uint16_t max_iterations)
 {
     for (int y = rect.tl.y; y < rect.br.y; y++)
     {
         for (int x = rect.tl.x; x < rect.br.x; x++)
         {
-            if (result[x + y * zoom->width] != max_iterations)
+            if ((uint16_t)result[x + y * zoom->width] != max_iterations)
             {
                 return false;
             }
@@ -39,7 +39,7 @@ static bool rect_is_black(const Rectangle rect, const uint16_t *result, const Zo
     return true;
 }
 
-static bool calculate_rect_border(const Rectangle rect, uint16_t *result, const Zoom *zoom, const uint16_t max_iterations)
+static bool calculate_rect_border(const Rectangle rect, double *result, const Zoom *zoom, const uint16_t max_iterations)
 {
     const Rectangle nw = (Rectangle){{rect.tl.x, rect.tl.y}, {rect.br.x, rect.tl.y + 1}};
     const Rectangle ne = (Rectangle){{rect.tl.x, rect.tl.y}, {rect.tl.x + 1, rect.br.y}};
@@ -81,7 +81,7 @@ static void split_rect(Queue *q, const Rectangle rect2)
     queue_push_back(q, se);
 }
 
-static void fill_black_rect(const Rectangle rect, uint16_t *result, const Zoom *zoom, const uint16_t max_iterations)
+static void fill_black_rect(const Rectangle rect, double *result, const Zoom *zoom, const uint16_t max_iterations)
 {
     for (int y = rect.tl.y; y < rect.br.y; y++)
     {
@@ -137,7 +137,7 @@ void *thread_work_four_split(void *threadwork)
     }
 }
 
-void calculate_mandelbrot_four_split(Zoom zoom, const uint16_t max_iterations, uint16_t *result, const uint16_t thread_count, const uint16_t square_size)
+void calculate_mandelbrot_four_split(Zoom zoom, const uint16_t max_iterations, double *result, const uint16_t thread_count, const uint16_t square_size)
 {
     Queue *q = queue_init();
 
@@ -165,7 +165,7 @@ void calculate_mandelbrot_four_split(Zoom zoom, const uint16_t max_iterations, u
     free(sem);
 }
 
-void calculate_mandelbrot_four_split_benchmark(const Zoom zoom, const uint16_t max_iterations, uint16_t *result, const uint16_t thread_count)
+void calculate_mandelbrot_four_split_benchmark(const Zoom zoom, const uint16_t max_iterations, double *result, const uint16_t thread_count)
 {
     printf("\n");
     float best_time = 9999999;
